nasc-empresa/requirements.txt
# Google AI SDK
google-genai==0.3.0
google-adk==0.3.0

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Environment and Configuration
python-dotenv==1.0.0

# HTTP Client
requests==2.31.0

# Data Validation
pydantic==2.5.0

# Authentication
python-jose[cryptography]==3.3.0

# Database (for session management)
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Logging
colorlog==6.8.0